Remove deprecated `GET /api/v1/search` API
Use `GET /api/v2/search` instead
